+++ src/main.lua
@@ -1,7 +1,5 @@
 local function Main()
-    -- require 'test.framework.behavior'
-    print(string.gmatch("烈焰-火焰強化", '.+'))
-    print(string.match("火焰強化", '-.+'))
+    require 'test.framework.behavior'
 end
 
 Main()